pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "robot"
version = "0.1.0"
description = "A context-aware terminal assistant"
authors = [
    {name = "Robot Dev", email = "robot@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all] >= 0.9",
    "openai>=1",
    "google-generativeai>=0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
]

[project.scripts]
robot = "robot.cli:main"
